requirements.txt
# Core Backend Dependencies
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.7.1
#sqlite3
sqlalchemy==2.0.29

# Authentication (Optional)
python-jose==3.3.0
passlib[bcrypt]==1.7.4

# AI/NLP Dependencies
transformers==4.41.0
sentence-transformers==2.7.0
torch==2.3.0
scikit-learn==1.4.2
nltk==3.8.1

# Testing
pytest==8.2.1
httpx==0.27.0

# Environment and Logging
python-dotenv==1.0.1
loguru==0.7.2

# Development Tools
black==24.4.2
isort==5.13.2

